+++ queue/testinit.go
@@ -1,0 +1,15 @@
+package queue
+
+var IsTest = false
+
+func SetIsTest() {
+	IsTest = true
+
+	Start()
+}
+
+func UnsetIsTest() {
+	IsTest = false
+
+	Close()
+}
