+++ queue/testinit_test.go
@@ -7,20 +7,6 @@ import (
 	"github.com/Ptt-official-app/go-openbbsmiddleware/utils"
 )
 
-var IsTest = false
-
-func SetIsTest() {
-	IsTest = true
-
-	Start()
-}
-
-func UnsetIsTest() {
-	IsTest = false
-
-	Close()
-}
-
 func setupTest() {
 	utils.SetIsTest()
